# docs/07_api/openapi.yaml
openapi: 3.0.3
info:
  title: WebCV Ung Vien Tuyen Dung API
  version: 1.4.0
  description: MVP API sau reliability/CAS/fingerprinted-idempotency lock. Parse create endpoints dùng Idempotency-Key, SHA-256 request fingerprint và internal recovery.
servers:
  - url: http://localhost:8000/api/v1
security:
  - BearerAuth: []
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      summary: Danh sách JD theo role
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Limit20'}
        - {name: keyword, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/JobStatus'}}
        - {name: parsing_status, in: query, schema: {$ref: '#/components/schemas/ParsingStatus'}}
      responses:
        '200': {description: Jobs, content: {application/json: {schema: {$ref: '#/components/schemas/PaginatedJobsResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
    post:
      tags: [Jobs]
      summary: HR tạo JD mới idempotently
      description: HR only; fixed UUIDv5 namespace and SHA-256 canonical create-payload fingerprint. Same key/same fingerprint returns same JD; same key/different fingerprint returns 409 IDEMPOTENCY_KEY_REUSED. Persist trước dispatch; dispatcher lỗi được recovery xử lý khi JD còn PENDING.
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/JobCreateRequest'}
      responses:
        '201': {description: Persisted/created; same key and same fingerprint returns same JD, content: {application/json: {schema: {$ref: '#/components/schemas/JobResponse'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/IdempotencyConflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /jobs/{id}:
    get:
      tags: [Jobs]
      summary: Chi tiết JD
      parameters: [{$ref: '#/components/parameters/UuidId'}]
      responses:
        '200': {description: Job, content: {application/json: {schema: {$ref: '#/components/schemas/JobResponse'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Jobs]
      summary: Cập nhật JD
      description: raw_content đổi => revision++ trước reparse, DRAFT/PENDING, invalidate Match; dispatch failure được PENDING recovery xử lý.
      parameters: [{$ref: '#/components/parameters/UuidId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/JobUpdateRequest'}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/JobResponse'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
    delete:
      tags: [Jobs]
      summary: Soft delete JD
      parameters: [{$ref: '#/components/parameters/UuidId'}]
      responses:
        '204': {description: Deleted}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /jobs/{id}/status:
    patch:
      tags: [Jobs]
      summary: Đổi trạng thái JD
      description: "Allowed: same-state, DRAFT->ACTIVE, ACTIVE->DRAFT|CLOSED, CLOSED->DRAFT|ACTIVE. DRAFT->CLOSED invalid."
      parameters: [{$ref: '#/components/parameters/UuidId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/JobStatusRequest'}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/JobResponse'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /jobs/{id}/criteria:
    get:
      tags: [Jobs]
      summary: Lấy criteria
      parameters: [{$ref: '#/components/parameters/UuidId'}]
      responses:
        '200': {description: Criteria, content: {application/json: {schema: {$ref: '#/components/schemas/JobCriteriaResponse'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Jobs]
      summary: Cập nhật criteria
      description: Direct mutation revision++ và invalidate Match generation/snapshots.
      parameters: [{$ref: '#/components/parameters/UuidId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/JobCriteriaRequest'}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/JobCriteriaResponse'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /jobs/{id}/weights:
    put:
      tags: [Jobs]
      summary: Cập nhật trọng số
      description: Chỉ Job current PARSED. Direct mutation revision++ và invalidate toàn bộ Match; recalculate chỉ quyết định best-effort dispatch sau commit. Persistence thành công luôn trả 200 dù publication lỗi.
      parameters: [{$ref: '#/components/parameters/UuidId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WeightsRequest'}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/JobResponse'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
...
components:
  securitySchemes:
    BearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    UuidId: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    Limit20: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID unique per logical create request. Stable UUIDv5 namespace derives resource id; same key is valid only when SHA-256 request fingerprint matches the persisted fingerprint.
      schema: {type: string, format: uuid}
  responses:
...
    Unauthorized: {description: Unauthorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Forbidden: {description: Forbidden, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    NotFound: {description: Not found or hidden, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
    IdempotencyConflict: {description: Same Idempotency-Key/deterministic resource id was reused with a different request fingerprint; error.code=IDEMPOTENCY_KEY_REUSED, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
    ValidationError: {description: Validation/precondition error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  schemas:
...
    ParsingStatus: {type: string, enum: [PENDING, PROCESSING, PARSED, FAILED]}
    JobStatus: {type: string, enum: [DRAFT, ACTIVE, CLOSED]}
...
    Meta:
      type: object
      required: [page, limit, total_items, total_pages]
      properties:
        page: {type: integer, minimum: 1}
        limit: {type: integer, minimum: 1}
        total_items: {type: integer, minimum: 0}
        total_pages: {type: integer, minimum: 0}
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: {type: boolean, enum: [false]}
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            details: {nullable: true}
...
    JobCreateRequest:
      type: object
      required: [title, job_level, raw_content]
      properties:
        title: {type: string, minLength: 1}
        job_level: {type: string, minLength: 1}
        location: {type: string, nullable: true}
        raw_content: {type: string, minLength: 1}
        w_skill: {type: number, minimum: 0, maximum: 1, default: 0.5}
        w_semantic: {type: number, minimum: 0, maximum: 1, default: 0.3}
        w_experience: {type: number, minimum: 0, maximum: 1, default: 0.2}
    JobUpdateRequest:
      type: object
      minProperties: 1
      properties:
        title: {type: string}
        job_level: {type: string}
        location: {type: string, nullable: true}
        raw_content: {type: string, minLength: 1}
    JobStatusRequest:
      type: object
      required: [status]
      properties:
        status: {$ref: '#/components/schemas/JobStatus'}
    Job:
      type: object
      required: [id, recruiter_id, revision, title, job_level, parsing_status, is_criteria_verified, w_skill, w_semantic, w_experience, status, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        recruiter_id: {type: string, format: uuid}
        revision: {type: integer, minimum: 1}
        title: {type: string}
        job_level: {type: string}
        location: {type: string, nullable: true}
        raw_content: {type: string}
        min_experience_years: {type: number, minimum: 0}
        education_requirement: {type: string, nullable: true}
        parsing_status: {$ref: '#/components/schemas/ParsingStatus'}
        is_criteria_verified: {type: boolean}
        w_skill: {type: number}
        w_semantic: {type: number}
        w_experience: {type: number}
        status: {$ref: '#/components/schemas/JobStatus'}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        parsed_at: {type: string, format: date-time, nullable: true}
    JobResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data: {$ref: '#/components/schemas/Job'}
    PaginatedJobsResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, enum: [true]}
        data: {type: array, items: {$ref: '#/components/schemas/Job'}}
        meta: {$ref: '#/components/schemas/Meta'}
    JobSkillCriterion:
      type: object
      required: [skill_id, importance]
      properties:
        skill_id: {type: integer}
        importance: {type: string, enum: [MANDATORY, OPTIONAL]}
        min_years_required: {type: number, minimum: 0, default: 0}
    JobCriteriaRequest:
      type: object
      required: [skills]
      properties:
        min_experience_years: {type: number, minimum: 0}
        education_requirement: {type: string, nullable: true}
        skills:
          type: array
          minItems: 1
          uniqueItems: true
          items: {$ref: '#/components/schemas/JobSkillCriterion'}
    JobCriteriaData:
      type: object
      required: [job_id, revision, is_criteria_verified, skills]
      properties:
        job_id: {type: string, format: uuid}
        revision: {type: integer, minimum: 1}
        min_experience_years: {type: number}
        education_requirement: {type: string, nullable: true}
        is_criteria_verified: {type: boolean}
        skills: {type: array, items: {$ref: '#/components/schemas/JobSkillCriterion'}}
    JobCriteriaResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data: {$ref: '#/components/schemas/JobCriteriaData'}
    WeightsRequest:
      type: object
      additionalProperties: false
      required: [w_skill, w_semantic, w_experience]
      properties:
        w_skill: {type: number, minimum: 0, maximum: 1}
        w_semantic: {type: number, minimum: 0, maximum: 1}
        w_experience: {type: number, minimum: 0, maximum: 1}
        recalculate: {type: boolean, default: false}
...